requirements.txt
Flask
Pillow
animegan  # Assurez-vous que AnimeGAN est installé et configuré correctement
werkzeug
numpy
tensorflow